requirements.txt
google-generativeai
neo4j
python-dotenv
requests
